# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Final Project: Tiketku",
    "description": "Platform ini merupakan tempat beli dan booking tiket penerbangan secara online, baik one way (sekali penerbangan) dan round trip (pulang pergi). Platform ini membuka dan menyediakan berbagai jenis penerbangan domestik dan mancanegara. Flow / alur platform dibebaskan berdasarkan kreasi tim final project. Terdapat 1 (satu) peran User yaitu login sebagai buyer. Buyer dapat melakukan pencarian, pemilihan, dan melihat histori pemesanan dan detail tiket.",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "brahmastabagus@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://be-tiketku-production.up.railway.app/api/v1",
      "description": "Production server"
    },
    {
      "url": "https://743b-103-162-237-17.ngrok-free.app/api/v1/",
      "description": "Stagging server"
    },
    {
      "url": "http://localhost:8080/api/v1/",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/user/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get user",
        "description": "Get user by id.",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success get users by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                }
              }
            }
          },
          "500": {
            "description": "Invalid input supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoData"
                }
              }
            }
          },
          "404": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res404"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Update user",
        "description": "Update user data.",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Updated user object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success Update user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Data dengan index 2 telah berhasil terupdate"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Invalid input supplied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "failed"
                    },
                    "message": {
                      "type": "string",
                      "example": "Terjadi kesalahan pada server"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res404"
                }
              }
            }
          },
          "400": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res400"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Delete user",
        "description": "Delete user by id.",
        "operationId": "deleteUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success delete user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Data dengan index 3 telah berhasil terhapus"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Invalid input supplied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "failed"
                    },
                    "message": {
                      "type": "string",
                      "example": "Terjadi kesalahan pada server"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res404"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      }
    },
...
    "/seat/{id}": {
      "get": {
        "tags": [
          "Seat"
        ],
        "summary": "Get seat",
        "description": "Get seat by id.",
        "operationId": "getSeatById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of seat to return",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success get seats by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Seats"
                }
              }
            }
          },
          "500": {
            "description": "Invalid input supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoData"
                }
              }
            }
          },
          "404": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res404"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Seat"
        ],
        "summary": "Update seat",
        "description": "Update seat data.",
        "operationId": "updateSeat",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of seat to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Updated seat object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Seat"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success Update seat",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Data dengan index 2 telah berhasil terupdate"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Invalid input supplied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "failed"
                    },
                    "message": {
                      "type": "string",
                      "example": "Terjadi kesalahan pada server"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res404"
                }
              }
            }
          },
          "400": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res400"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Seat"
        ],
        "summary": "Delete seat",
        "description": "Delete seat by id.",
        "operationId": "deleteSeatById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of seat to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success delete seat",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Data dengan index 3 telah berhasil terhapus"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Invalid input supplied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "failed"
                    },
                    "message": {
                      "type": "string",
                      "example": "Terjadi kesalahan pada server"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res404"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "Admin11#"
          },
          "first_name": {
            "type": "string",
            "example": "saya",
            "description": "Nama Lengkap"
          },
          "last_name": {
            "type": "string",
            "example": "user",
            "description": "Nama Keluarga"
          },
          "phone_number": {
            "type": "integer",
            "example": "08123456789"
          },
          "role": {
            "type": "string",
            "example": "admin",
            "enum": [
              "admin",
              "user"
            ]
          }
        },
        "required": [
          "email",
          "password",
          "first_name",
          "phone_number"
        ]
      },
      "Users": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "email": {
            "type": "string",
            "example": "user@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "Admin11#"
          },
          "first_name": {
            "type": "string",
            "example": "saya",
            "description": "Nama Lengkap"
          },
          "last_name": {
            "type": "string",
            "example": "user",
            "description": "Nama Keluarga"
          },
          "phone_number": {
            "type": "integer",
            "example": "08123456789"
          },
          "google_id": {
            "type": "integer",
            "example": "1"
          },
          "role": {
            "type": "string",
            "example": "admin",
            "enum": [
              "admin",
              "user"
            ]
          },
          "otp": {
            "type": "integer",
            "example": "12345"
          },
          "createdAt": {
            "type": "string",
            "example": "2023-05-04T14:20:27.537Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2023-05-04T14:20:27.537Z"
          }
        }
      },
...
      "Seat": {
        "type": "object",
        "properties": {
          "seat_number": {
            "type": "string",
            "example": "A1"
          },
          "ticket_id": {
            "type": "integer",
            "example": 1
          },
          "flight_id": {
            "type": "integer",
            "example": 2
          },
          "status": {
            "type": "string",
            "example": "Available",
            "enum": [
              "Available",
              "Unavailable"
            ]
          }
        },
        "required": [
          "seat_number",
          "ticket_id",
          "flight_id",
          "status"
        ]
      },
      "Seats": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "seat_number": {
            "type": "string",
            "example": "A1"
          },
          "ticket_id": {
            "type": "integer",
            "example": 1
          },
          "flight_id": {
            "type": "integer",
            "example": 2
          },
          "status": {
            "type": "string",
            "example": "Available",
            "enum": [
              "Available",
              "Unavailable"
            ]
          },
          "createdAt": {
            "type": "string",
            "example": "2023-05-04T14:20:27.537Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2023-05-04T14:20:27.537Z"
          }
        }
      },
...
      "NoData": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "failed"
          },
          "data": {
            "type": "array",
            "example": [ ]
          }
        }
      },
...
      "Res404": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "failed"
          },
          "message": {
            "type": "string",
            "example": "URL tidak ditemukan"
          }
        }
      },
      "Res400": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "failed"
          },
          "message": {
            "type": "string",
            "example": "\"name\" is required"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}